\name{cancerRates}
\alias{cancerRates}
\title{Breast Cancer Dataset}
\description{
Breast cancer rate for Swedish females between age of 40-50,
courtesy of Professor Y. Pawitan.
A data frame with 99 observations on the following 2 variables.
\code{Age}: Age of women having the disease.
\code{Incidence}: Number of women of that age.
}

