% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{changes}
\alias{changes}
\title{View Changes in Files under Version Control.}
\usage{
changes(files = NULL)
}
\arguments{
\item{files}{an optional character vector of files for which you want to see
the changes}
}
\value{
number of changes
}
\description{
TODO Describe this better.
}
