% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo.R
\name{create_repo}
\alias{create_repo}
\title{Create a Repository for Version Control}
\usage{
create_repo(path = getwd(), add_dirs = TRUE, change_dir = TRUE)
}
\arguments{
\item{path}{TODO}

\item{add_dirs}{TODO}

\item{change_dir}{TODO}
}
\description{
TODO: Describe this better.
}
