% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.decon.R
\name{plot.history}
\alias{plot.history}
\title{Sets default plotting for history log}
\usage{
\method{plot}{decon}(x, ...)
}
\arguments{
\item{x}{history object as generated by history function}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
This function sets up the default plotting method
for outputs from deconvolve function
}
