% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.timeline.R
\name{plot.timeline}
\alias{plot.timeline}
\title{Sets default plotting for timeline}
\usage{
\method{plot}{timeline}(x, records = 6, ...)
}
\arguments{
\item{x}{timeline object as generated by timeline function}

\item{records}{list the number of records you'd like to show}

\item{...}{other options passed to plot}
}
\value{
plot
}
\description{
This function sets up the default plotting method
for outputs from timeline function
}
