% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.history.R
\name{print.history}
\alias{print.history}
\title{Sets default print method for history logs}
\usage{
\method{print}{history}(x, ...)
}
\arguments{
\item{x}{history object as generated by history function}

\item{...}{other options passed to print}
}
\value{
print history
}
\description{
This function sets up the default print method
for outputs of the history function
}
