% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.timeline.R
\name{print.timeline}
\alias{print.timeline}
\title{Sets default print method for timeline}
\usage{
\method{print}{timeline}(x, ...)
}
\arguments{
\item{x}{timeline object as generated by timeline function}

\item{...}{other options passed to print}
}
\value{
print timeline
}
\description{
This function sets up the default print method
for outputs of the timeline function
}
