% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminder.R
\name{reminder_delay}
\alias{reminder_delay}
\title{Set the reminder delay in minutes for commit reminder messages}
\usage{
reminder_delay(minutes)
}
\arguments{
\item{minutes}{delay in minutes}
}
\value{
nothing
}
\description{
TODO: Describe this better.
}
