% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reminder.R
\name{schedule_reminder}
\alias{schedule_reminder}
\title{Schedule a reminder}
\usage{
schedule_reminder()
}
\description{
TODO: This needs to be called at the tope of every stow function, as well as in the .onLoad of the package
}
\examples{
TODO
}
