% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{scrub}
\alias{scrub}
\title{Remove all unstaged changes}
\usage{
scrub()
}
\description{
Undo all of the changes you made since the last time you did \code{record()}.
\emph{WARNING: this will delete and change files and can't be undone!}
}
