% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starlogs.R
\name{starlog}
\alias{starlog}
\title{Show repo history in starlogs}
\usage{
starlog()
}
\description{
Display your past record messages at starlogs.net. This requires internet
access and only works for version control projects that have an online copy
}
