% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{unignore}
\alias{unignore}
\title{Remove files or folders to list of ignored items}
\usage{
unignore(files)
}
\arguments{
\item{files}{vector or list files to retrive}
}
\value{
TODO
}
\description{
TODO Describe this better.
}
