\name{cancerRates}
\alias{cancerRates}
\docType{data}
\title{
  Breast Cancer Rate
}
\description{
  Breast cancer rate for Swedish females between age of 40-50.
}
\usage{data("bcn")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{a numeric vector}
    \item{\code{Incidence}}{a numeric vector}
  }
}

\source{
  https://doi.org/10.1002/0470011815.b2a12011
}

