\name{cancerRates}
\alias{cancerRates}
\docType{data}
\title{
  Breast Cancer Dataset
}
\description{
  Breast cancer rate for Swedish females between age of 40-50,
  courtesy of Professor Y. Pawitan.
}
\usage{data("cancerRates")}
\format{
  A data frame with 99 observations on the following 2 variables.
  \describe{
    \item{\code{Age}}{Age of women having the disease.}
    \item{\code{Incidence}}{Number of women of that age.}
  }
}

\source{
  https://doi.org/10.1002/0470011815.b2a12011
}

