% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo.R
\name{create_repo}
\alias{create_repo}
\title{Create a Repository for Version Control}
\usage{
create_repo(path = getwd(), add_structure = TRUE, change_wd = TRUE,
  remind_me_after = 60)
}
\arguments{
\item{path}{TODO}

\item{add_structure}{TODO}

\item{change_wd}{TODO}

\item{remind_me_after}{set reminder delay in minutes, zero disables reminders}
}
\description{
TODO: Describe this better.
}
\details{
See \code{\link{remind_me}()} for more information on reminders
}
