% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repo.R
\name{create_repo}
\alias{create_repo}
\title{Create a Repository for Version Control}
\usage{
create_repo(path = getwd(), add_structure = TRUE, change_wd = TRUE,
  reminders = 30)
}
\arguments{
\item{path}{TODO}

\item{add_structure}{TODO}

\item{change_wd}{TODO}

\item{reminders}{set reminder delay in minutes, zero disables reminders}
}
\description{
TODO: Describe this better.
}
