% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_to.R
\name{go_to}
\alias{go_to}
\title{Go to a Specified Record in the Version Control Timeline}
\usage{
go_to(number)
}
\arguments{
\item{number}{the record number to go to}
}
\description{
TODO Describe this better.
}
