% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_to.R
\name{go_to}
\alias{go_to}
\title{Go to a Specified Snapshot in the Version Control Timeline}
\usage{
go_to(sha)
}
\arguments{
\item{sha}{The snapshot to go to.}
}
\value{
TODO
}
\description{
TODO Describe this better.
}
