% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore.R
\name{ignore}
\alias{ignore}
\title{Add files or folders to list of ignored items}
\usage{
ignore(files)
}
\arguments{
\item{files}{vector or list files to retrive}
}
\value{
TODO
}
\description{
TODO Describe this better.
}
