% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record Changes in Files under Version Control.}
\usage{
record(message)
}
\arguments{
\item{message}{character: the message to be added to the version control
checkpoint.}
}
\value{
the git2r commit object (S4). TODO: maybe return something nicer.
}
\description{
TODO Describe this better.
}
