% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve code from a past commit.}
\usage{
retrieve(sha, files = NULL)
}
\arguments{
\item{sha}{The sha to retrieve}

\item{files}{vector or list files to retrive TODO: add this}
}
\value{
TODO
}
\description{
If the argument files is NULL will reteive state of entire repo
}
\details{
TODO Describe this better.
}
